# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TrailTeam team management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List events
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: eventType
          in: query
          schema:
            type: string
            enum: [practice, race, social, volunteer, other]
        - name: podId
          in: query
          schema:
            type: string
        - name: archived
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EventWithDetails"
    post:
      operationId: createEvent
      tags: [events]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventBody"
      responses:
        "201":
          description: Created event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventWithDetails"
  /events/batch:
    post:
      operationId: batchCreateEvents
      tags: [events]
      summary: Batch-create multiple events (season builder)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchCreateEventsBody"
      responses:
        "201":
          description: Batch create result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BatchCreateEventsResult"
  /series/{seriesId}:
    delete:
      operationId: deleteSeries
      tags: [events]
      summary: Delete all future events in a series (coach/admin only)
      parameters:
        - name: seriesId
          in: path
          required: true
          schema:
            type: string
        - name: fromDate
          in: query
          description: Delete events on or after this date (ISO). Defaults to now.
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Deletion result
          content:
            application/json:
              schema:
                type: object
                properties:
                  deleted:
                    type: integer
                required: [deleted]
  /series/{seriesId}/reschedule:
    patch:
      operationId: rescheduleSeries
      tags: [events]
      summary: Shift all future events in a series by N days (coach/admin only)
      parameters:
        - name: seriesId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                shiftDays:
                  type: integer
                  description: Number of days to shift (positive = forward, negative = backward)
                fromDate:
                  type: string
                  format: date
                  description: Only reschedule events on or after this date. Defaults to now.
              required: [shiftDays]
      responses:
        "200":
          description: Reschedule result
          content:
            application/json:
              schema:
                type: object
                properties:
                  rescheduled:
                    type: integer
                required: [rescheduled]
  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event with details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventWithDetails"
    patch:
      operationId: updateEvent
      tags: [events]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEventBody"
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventWithDetails"
    delete:
      operationId: deleteEvent
      tags: [events]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /events/{id}/rsvp:
    post:
      operationId: rsvpEvent
      tags: [rsvps]
      summary: RSVP to an event (or update existing RSVP)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RsvpBody"
      responses:
        "200":
          description: Updated RSVP
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventRsvp"
  /events/{id}/rsvps:
    get:
      operationId: listEventRsvps
      tags: [rsvps]
      summary: List RSVPs for an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of RSVPs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EventRsvpWithUser"
...
  /events/{id}/attachments:
    post:
      operationId: addEventAttachment
      tags: [events]
      summary: Add a file attachment (e.g. race PDF flyer) to an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddAttachmentBody"
      responses:
        "201":
          description: Attachment added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventAttachment"

  # ─── CARPOOLS ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    UserNotificationPreferences:
      type: object
      properties:
        practiceReminders:
          type: boolean
        coachMessages:
          type: boolean
        carpoolUpdates:
          type: boolean
        eventReminders:
          type: boolean
        rosterUpdates:
          type: boolean
        boardReplies:
          type: boolean
      required:
        - practiceReminders
        - coachMessages
        - carpoolUpdates
        - eventReminders
        - rosterUpdates
        - boardReplies
    User:
      type: object
      properties:
        id:
          type: integer
        householdId:
          type: integer
          nullable: true
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        role:
          type: string
          enum: [super_admin, coach, parent, student]
        podId:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        isActive:
          type: boolean
        gender:
          type: string
          nullable: true
        grade:
          type: integer
          nullable: true
        coachCertLevel:
          type: string
          nullable: true
        notificationsEnabled:
          type: boolean
        emailNotifications:
          type: boolean
        smsNotifications:
          type: boolean
        pushNotifications:
          type: boolean
        defaultCarpoolSeats:
          type: integer
          nullable: true
        defaultCarpoolTrays:
          type: integer
          nullable: true
        notificationPreferences:
          $ref: '#/components/schemas/UserNotificationPreferences'
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - firstName
        - lastName
        - email
        - role
        - isActive
        - notificationsEnabled
        - emailNotifications
        - smsNotifications
        - pushNotifications
        - createdAt
...
    Event:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        eventType:
          type: string
          enum: [practice, race, social, volunteer, other]
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
          nullable: true
        trailheadId:
          type: integer
          nullable: true
        locationOverride:
          type: string
          nullable: true
        googleMapsUrlOverride:
          type: string
          nullable: true
        podIds:
          type: array
          items:
            type: string
          nullable: true
        isAllTeam:
          type: boolean
        rsvpDeadline:
          type: string
          format: date-time
          nullable: true
        volunteerSlotsNeeded:
          type: integer
        iCalUid:
          type: string
        isArchived:
          type: boolean
        seriesId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - eventType
        - startTime
        - isAllTeam
        - volunteerSlotsNeeded
        - iCalUid
        - isArchived
        - createdAt
    HouseholdMemberRsvp:
      type: object
      properties:
        userId:
          type: integer
        firstName:
          type: string
        isMe:
          type: boolean
        status:
          type: string
          enum: [attending, not_attending, maybe]
          nullable: true
      required: [userId, firstName, isMe]
    EventWithDetails:
      allOf:
        - $ref: "#/components/schemas/Event"
        - type: object
          properties:
            trailhead:
              $ref: "#/components/schemas/Trailhead"
              nullable: true
            myRsvp:
              type: string
              enum: [attending, not_attending, maybe]
              nullable: true
            householdRsvps:
              type: array
              nullable: true
              items:
                $ref: "#/components/schemas/HouseholdMemberRsvp"
            rsvpCounts:
              $ref: "#/components/schemas/RsvpCounts"
            volunteerCount:
              type: integer
            carpoolSpotsAvailable:
              type: integer
            attachments:
              type: array
              items:
                $ref: "#/components/schemas/EventAttachment"
          required: [rsvpCounts, volunteerCount, carpoolSpotsAvailable, attachments]
    RsvpCounts:
      type: object
      properties:
        attending:
          type: integer
        notAttending:
          type: integer
        maybe:
          type: integer
      required: [attending, notAttending, maybe]
    EventAttachment:
      type: object
      properties:
        id:
          type: integer
        eventId:
          type: integer
        label:
          type: string
        objectPath:
          type: string
        mimeType:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, eventId, label, objectPath, createdAt]
    AddAttachmentBody:
      type: object
      properties:
        label:
          type: string
        objectPath:
          type: string
        mimeType:
          type: string
      required: [label, objectPath]
    CreateEventBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        eventType:
          type: string
          enum: [practice, race, social, volunteer, other]
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        trailheadId:
          type: integer
        locationOverride:
          type: string
        googleMapsUrlOverride:
          type: string
        podIds:
          type: array
          items:
            type: string
        isAllTeam:
          type: boolean
        rsvpDeadline:
          type: string
          format: date-time
        volunteerSlotsNeeded:
          type: integer
        seriesId:
          type: string
          nullable: true
      required: [title, eventType, startTime]
    BatchCreateEventsBody:
      type: object
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/CreateEventBody"
        seriesId:
          type: string
          description: Optional shared series ID to group these events
      required: [events]
    BatchCreateEventsResult:
      type: object
      properties:
        created:
          type: integer
          description: Number of events created
        ids:
          type: array
          items:
            type: integer
          description: IDs of the newly created events
      required: [created, ids]
    UpdateEventBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        eventType:
          type: string
          enum: [practice, race, social, volunteer, other]
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        trailheadId:
          type: integer
        locationOverride:
          type: string
        podIds:
          type: array
          items:
            type: string
        isAllTeam:
          type: boolean
        rsvpDeadline:
          type: string
          format: date-time
        volunteerSlotsNeeded:
          type: integer
        isArchived:
          type: boolean
        seriesId:
          type: string
          nullable: true

    # ─── RSVP ─────────────────────────────────────────────────────────────────
    RsvpBody:
      type: object
      properties:
        status:
          type: string
          enum: [attending, not_attending, maybe]
        userIds:
          type: array
          items:
            type: integer
          description: RSVP for specific household members (defaults to self)
      required: [status]
    EventRsvp:
      type: object
      properties:
        id:
          type: integer
        eventId:
          type: integer
        userId:
          type: integer
        status:
          type: string
          enum: [attending, not_attending, maybe]
        respondedAt:
          type: string
          format: date-time
      required: [id, eventId, userId, status, respondedAt]
    EventRsvpWithUser:
      allOf:
        - $ref: "#/components/schemas/EventRsvp"
        - type: object
          properties:
            user:
              $ref: "#/components/schemas/User"
          required: [user]
...
    Trailhead:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        address:
          type: string
          nullable: true
        googleMapsUrl:
          type: string
          nullable: true
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        notes:
          type: string
          nullable: true
        photoObjectPath:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, name, createdAt]
...